[SPARK-17269][SQL] Move finish analysis optimization stage into its own file
